refactor: split repos into subpackages
moved each repo to its own subpackage for cleaner imports
switched to pgx struct scanning instead of manual scan calls
changed daily_target from float to int across all layers

+++ internal/usecase/habit/dto.go
@@ -10,7 +10,7 @@ type CreateInput struct {
 	UserID      uuid.UUID
 	Name        string
 	ExpReward   int64
-	DailyTarget float64
+	DailyTarget int64
 	Unit        string
 }
 
@@ -31,7 +31,7 @@ type UpdateInput struct {
 	UserID      uuid.UUID
 	Name        string
 	ExpReward   int64
-	DailyTarget float64
+	DailyTarget int64
 	Unit        string
 }
